game changes to main

+++ Starship_Elemental/objects/obj_ctrl_pause/Step_0.gml
@@ -18,24 +18,41 @@ if instance_exists(obj_smartbomb_effect_ice) return;
 var start = Gamepad_start();
 
 if keyboard_check_pressed(vk_escape) || (start){
+	alarm[1] = GAMESPEED * 3;
+	exiting = true;
 	if (can_pause){
 		initAudio();
 		global.GAMEPaused = !global.GAMEPaused;	
 		playSound(SND_SFX_PAUSE);
-		if !(file_exists(pause_sprite)){
-			pause_png = (working_directory + "\\ScreenShots\\Screen_pause.png");
-			screen_save(pause_png);
+		
+		if !(global.IS_GX_EXPORT){
+			if !(file_exists(pause_sprite)){
+				pause_png = (working_directory + "\\ScreenShots\\Screen_pause.png");
+				screen_save(pause_png);
 	
-			pause_sprite = sprite_add(pause_png,1,false,false,0,0);
+				pause_sprite = sprite_add(pause_png,1,false,false,0,0);
+			}
+		}else{
+			if (global.GXC_Pause_Surface == -4){
+				global.GXC_Pause_Surface = surface_create(display_get_gui_width(),display_get_gui_height());
+				surface_copy(global.GXC_Pause_Surface,0,0,application_surface);
+			}
 		}
 		alarm[0] = GAMESPEED;
 		can_pause = false;
 	}
 }
 
+if (keyboard_check_released(vk_escape) || (start)){
+	exiting = false;	
+}
+
 if (global.GAMEPaused){
 	
 	instance_deactivate_all(true);
+	
+	//fixes audiobug
+	//fhAudioActivate();
 	
 	//keeps clouds moving
 	instance_activate_object(obj_clouds);
@@ -44,10 +61,19 @@ if (global.GAMEPaused){
 	instance_activate_object(obj_enemy_wave_spawner);
 	
 }else{
-	if (pause_sprite != -4){
-		if (file_exists(pause_sprite)){
-			sprite_delete(pause_sprite);
-			file_delete(pause_png);
+	if !(global.IS_GX_EXPORT){
+		if (pause_sprite != -4){
+			if (file_exists(pause_sprite)){
+				sprite_delete(pause_sprite);
+				file_delete(pause_png);
+			}
+		}
+	}else{
+		if (global.GXC_Pause_Surface != -4){
+			if (surface_exists(global.GXC_Pause_Surface)){
+				surface_free(global.GXC_Pause_Surface);
+			}
+			global.GXC_Pause_Surface = -4;
 		}
 	}
